refactor(sns): move configuration types into sns package
